Organize scripts and python code
