--- hacking_testing/compile_proto.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-protoc -I=../proto/ --python_out=proto_python/ --cpp_out=../src ../proto/*.proto
