--- hacking_testing/record_client.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-exec bin/minetest --name MinetestAgent --password whyisthisnecessary --address 0.0.0.0 --port 30000 --go --client-address "tcp://*:5555" --record --noresizing --cursor-image "cursors/mouse_cursor_white_16x16.png" --config hacking_testing/minetest.conf
