--- hacking_testing/server.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-exec bin/minetest --server --world newworld --gameid minetest --config hacking_testing/minetest.conf
