+++ scripts/compile_proto.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+protoc -I=../proto/ --python_out=../minetests/proto --cpp_out=../src ../proto/*.proto
