+++ scripts/data_recorder.py
@@ -1,7 +1,7 @@
 import os
 
 import zmq
-from minetest_env import unpack_pb_obs
+from minetests.utils import unpack_pb_obs
 
 
 class DataRecorder:
