+++ scripts/dumb_client.sh
