+++ scripts/minetest.conf
