+++ scripts/server.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+exec bin/minetest \
+    --server \
+    --world newworld \
+    --gameid minetest \
+    --config scripts/minetest.conf
